+++ app/views/layouts/application.html.erb
@@ -10,12 +10,8 @@
     <% default_page_title = "#{params[:controller]} / #{params[:action]}" %>
     <title><%= @page_title || default_page_title %> - Hourglass</title>
 
-    <%= stylesheet_pack_tag 'application' %>
-    <%= stylesheet_pack_tag 'bootstrap' %>
-
-    <%= javascript_pack_tag 'application' %>
-    <%= javascript_pack_tag 'bootstrap' %>
-    <%= javascript_pack_tag 'vendor' %>
+    <%= stylesheet_pack_tag 'application', 'bootstrap' %>
+    <%= javascript_pack_tag 'application', 'bootstrap' %>
 
     <%= favicon_pack_tag 'site-icon.png' %>
   </head>
