+++ relay.config.js
@@ -1,8 +1,8 @@
 module.exports = {
   // ...
   // Configuration options accepted by the `relay-compiler` command-line tool and `babel-plugin-relay`.
-  src: './app/javascript/components',
-  schema: './app/javascript/relay/data/schema.json',
+  src: './app/packs/components',
+  schema: './app/packs/relay/data/schema.json',
   exclude: ['**/node_modules/**', '**/__mocks__/**', '**/__generated__/**'],
   extensions: ['ts', 'tsx'],
   language: 'typescript',
